requirements.txt
docx
pillow
pyocr
pdf2image
pdfminer.six
openpyxl
numpy
